backend/audio_processing_service/requirements.txt
fastapi==0.112.0
uvicorn==0.15.0
pydantic==2.8.2
numpy==2.0.1
scipy==1.13.1
pydub==0.25.1
openai==1.42.0
boto3==1.35.9
aioredis==2.0.1
librosa==0.10.2.post1
pyannote-audio==3.3.1
soundfile==0.12.1



